requirements.txt
streamlit==1.31.0
pytesseract==0.3.10
Pillow==10.2.0
watchdog==3.0.0
